add messagerie system

+++ absence/src/main/java/miaad/rh/absence/service/impl/AbsenceServiceImpl.java
@@ -252,7 +252,9 @@ public class AbsenceServiceImpl implements AbsenceService {
         List<Absence> absences = absenceRepository.findByColaborateurId(id);
 
         for (Absence absence : absences) {
-            absenceRepository.delete(absence);
+            if(absence.isEmployee()){
+                absenceRepository.delete(absence);
+            }
         }
     }
